manifests: Drop etcd metrics from controller-manager targets
The etcd metrics are only accidentally intrumented. The reason they are
exposed by the controller-manager is because it uses the global metrics
registry and imports the etcd packages that instrument these metrics
that are really only meant for etcd servers.

These are dropped here, as their presence influences etcd dashboard's
correctness.