Respecting resource limits and requests during build.
This patch restores the functionality of restricting how much resources
a build can consume. Some extra test cases got added to validate the
functionality. To fix the problem function ConvertJSON() had to be
adapted to support maps also.

Tests added for ConvertJSON() function were designed with the old
behavior in mind as this function used to work only for slices fail
with anything else.