migrate operand ns from CVO
Recovery testing shows that operators that delegate operand ns creation to CVO
take too long (several minutes, at least) to recover after operand ns is deleted.
If other COs depend on health of the operand deployment, this sets a chain reaction
in the cluster.  On the other hand, if ns creation is controlled by the operator,
recovery happens immediately.