Bug 1826183: Sync proxy CA to openshift-controller-manager
Previously, builds mounted the global CA trust bundle as the PEM-encoded TLS
trust bundle. This trust bundle is generated by the Network Operator, which
validates any user-provided CA trust bundle and merges it with the default trus
bundle provided by RHCOS. Using the global CA is problematic because maven and
other JVM-based processes establish trust using `keytool` formatted trust
bundles.

To generate trust bundles in the `keytool` format (as well as other formats used
by RHEL), builds will run `update-ca-trust extract` on startup. To merge the
trust bundle correctly, the build controller needs to mount any user-provided CA
in a well-known location. The openshift-builder process is then responsible for
copying the CA to /etc/ca-trust/source/anchors prior to running
`update-ca-trust extract`.

To achieve this, the operator is being enhanced with the following:

* Added a separate controller for observing the user CA config.
* Added a ResourceSync controller to copy the user CA to the
  openshift-controller-manager namespace.
* Updated starter to launch the new controllers.

The operator will continue to create the `openshift-global-ca` ConfigMap to
ensure that the openshift controller manager does not have an outage during
upgrades. This can be removed in a subsequent release.