Allow ingress-to-route to update ingress status
Add a new
"system:openshift:openshift-controller-manager:ingress-to-route-controller"
cluster role and cluster role binding for the ingress-to-route controller.
The new cluster role is based on the
"system:openshift:controller:ingress-to-route-controller" cluster role from
the bootstrap policy, with the following changes:

* The new role references the ingresses resource in the networking.k8s.io
group, instead of using the ingresses resource in the extensions group,
which is deprecated.

* The new role allow updating the ingresses/status subresource.

This commit is related to bug 1820075.

https://bugzilla.redhat.com/show_bug.cgi?id=1820075

* bindata/v3.11.0/openshift-controller-manager/ingress-to-route-controller-clusterrole.yaml:
New file, based on system:openshift:controller:ingress-to-route-controller
from the bootstrap cluster roles.  Replace the extensions API group with
the networking.k8s.io API group for ingresses.  Allow the ingress-to-route
controller to update the status of ingresses.
* bindata/v3.11.0/openshift-controller-manager/ingress-to-route-controller-clusterrolebinding.yaml:
New file, based on system:openshift:controller:ingress-to-route-controller
from the bootstrap cluster role bindings.
* pkg/operator/sync_openshiftcontrollermanager_v311_00.go
(syncOpenShiftControllerManager_v311_00_to_latest): Add the new resources.
* pkg/operator/v311_00_assets/bindata.go: Regenerate.