Always sort disabled controller list
Currently, controller list is unsorted and any order change in that list
(even though the values are not changed) causes new OCM roll out.

This is happening only when the capabitilies are disabled because
if there is no disabled capability, controller list has only 1 item.

This commit always sorts this list for consistency and prevent
unnecessary rollouts.