OCPBUGS-52193: remove finalizer from legacy registry token
The legacy token controller manages the `openshift.io/legacy-token`
finalizer on docker config secrets. This controller runs as part of the
operand for this operator, except when the registry is in a "Removed"
state. When in the "Removed" state, the legacy token controller does not
run and therefore cannot remove the finalizer, preventing the deletion
of these secrets.

A routine within the operator is responsible for deleting the registry
secrets when in the Removed state. This cleanup routine must also remove
the finalizer to allow successful deletion of the secrets.

This commit addresses the issue by:

1. Ensuring the cleanup routine in the ocm-o removes the
   `openshift.io/legacy-token` finalizer from the docker config secrets.
2. Starting the cleanup routine only when the operand is fully deployed
   to ensure all necessary components are available.